prm graph experiments and file cleanup
