+++ mav_path_smoothing/src/polynomial_smoother.cpp
@@ -129,7 +129,9 @@ bool PolynomialSmoother::getTrajectoryBetweenWaypoints(
         break;
       }
     }
-    ROS_INFO("[SPLIT SMOOTHING] Added %d additional vertices.", num_added);
+    if (verbose_) {
+      ROS_INFO("[SPLIT SMOOTHING] Added %d additional vertices.", num_added);
+    }
 
     // If we had to do this, let's scale the times back to make sure we're
     // still within constraints.
@@ -142,8 +144,10 @@ bool PolynomialSmoother::getTrajectoryBetweenWaypoints(
   double v_max, a_max;
   trajectory->computeMaxVelocityAndAcceleration(&v_max, &a_max);
 
-  ROS_INFO("[SMOOTHING] V max/limit: %f/%f, A max/limit: %f/%f", v_max,
-           constraints_.v_max, a_max, constraints_.a_max);
+  if (verbose_) {
+    ROS_INFO("[SMOOTHING] V max/limit: %f/%f, A max/limit: %f/%f", v_max,
+             constraints_.v_max, a_max, constraints_.a_max);
+  }
 
   return true;
 }
